+++ src/logger.ts
@@ -28,7 +28,7 @@ type Logger = {
 	clear(),
 };
 
-const LOG_SIZE = 26;
+const LOG_SIZE = 16;
 
 function loggerInit(gfx: Gfx, assets: Assets, conf: LoggerConf = {
 	max: 8,